+++ solfees-be/src/metrics.rs
@@ -102,10 +102,14 @@ pub mod solfees_be {
             grpc_geyser::CommitmentLevel,
             rpc_solana::{RpcRequestsStats, SolanaRpcMode},
         },
-        http::StatusCode,
+        http::{HeaderMap, StatusCode},
         prometheus::{HistogramOpts, HistogramVec, IntCounterVec, IntGauge, IntGaugeVec, Opts},
         solana_sdk::clock::Slot,
-        std::{borrow::Cow, time::Duration},
+        std::{
+            borrow::Cow,
+            sync::Arc,
+            time::{Duration, Instant},
+        },
     };
 
     lazy_static::lazy_static! {
@@ -137,6 +141,11 @@ pub mod solfees_be {
             Opts::new("websockets_alive_total", "Total number of alive WebSocket connections"),
             &["api"]
         ).unwrap();
+
+        static ref CLIENT_USAGE_CPU_TOTAL: IntCounterVec = IntCounterVec::new(
+            Opts::new("client_usage_cpu_total", "Total number of CPU usage in nanoseconds"),
+            &["client_id", "subscription_id"]
+        ).unwrap();
     }
 
     pub fn init() {
@@ -147,6 +156,7 @@ pub mod solfees_be {
         register!(REQUESTS_CALLS_TOTAL);
         register!(REQUESTS_QUEUE_SIZE);
         register!(WEBSOCKETS_ALIVE_TOTAL);
+        register!(CLIENT_USAGE_CPU_TOTAL);
     }
 
     pub fn set_slot(commitment: CommitmentLevel, slot: Slot) {
@@ -179,16 +189,16 @@ pub mod solfees_be {
             .inc_by(stats.latest_blockhash);
         REQUESTS_CALLS_TOTAL
             .with_label_values(&[api.as_str(), "get_leader_schedule"])
-            .inc_by(stats.latest_blockhash);
+            .inc_by(stats.leader_schedule);
         REQUESTS_CALLS_TOTAL
             .with_label_values(&[api.as_str(), "get_recent_prioritization_fees"])
-            .inc_by(stats.latest_blockhash);
+            .inc_by(stats.recent_prioritization_fees);
         REQUESTS_CALLS_TOTAL
             .with_label_values(&[api.as_str(), "get_slot"])
-            .inc_by(stats.latest_blockhash);
+            .inc_by(stats.slot);
         REQUESTS_CALLS_TOTAL
             .with_label_values(&[api.as_str(), "get_version"])
-            .inc_by(stats.latest_blockhash);
+            .inc_by(stats.version);
     }
 
     pub fn requests_queue_size_inc() {
@@ -215,5 +225,87 @@ pub mod solfees_be {
             _ => return,
         };
         WEBSOCKETS_ALIVE_TOTAL.with_label_values(&[api]).dec()
+    }
+
+    #[derive(Debug)]
+    struct ClientIdInner {
+        client_id: String,
+        subsription_id: String,
+    }
+
+    #[derive(Debug, Clone)]
+    pub struct ClientId {
+        inner: Arc<ClientIdInner>,
+    }
+
+    impl ClientId {
+        pub fn new(headers: &HeaderMap) -> Self {
+            Self {
+                inner: Arc::new(ClientIdInner {
+                    client_id: Self::get(headers, "x-client-id"),
+                    subsription_id: Self::get(headers, "x-subscription-id"),
+                }),
+            }
+        }
+
+        fn get(headers: &HeaderMap, key: &str) -> String {
+            headers
+                .get(key)
+                .and_then(|id| id.to_str().ok())
+                .map(String::from)
+                .unwrap_or_default()
+        }
+
+        pub fn start_timer(&self) -> ClientIdTimer {
+            ClientIdTimer::new(self)
+        }
+
+        fn observe(&self, nanos: u64) {
+            CLIENT_USAGE_CPU_TOTAL
+                .with_label_values(&[&self.inner.client_id, &self.inner.subsription_id])
+                .inc_by(nanos);
+        }
+    }
+
+    #[derive(Debug)]
+    pub struct ClientIdTimer<'a> {
+        client_id: &'a ClientId,
+        start: Instant,
+        observed: bool,
+    }
+
+    impl<'a> Drop for ClientIdTimer<'a> {
+        fn drop(&mut self) {
+            if !self.observed {
+                self.observe(true);
+            }
+        }
+    }
+
+    impl<'a> ClientIdTimer<'a> {
+        fn new(client_id: &'a ClientId) -> Self {
+            Self {
+                client_id,
+                start: Instant::now(),
+                observed: false,
+            }
+        }
+
+        fn observe(&mut self, record: bool) {
+            self.observed = true;
+            if record {
+                let elapsed = Instant::now().saturating_duration_since(self.start);
+                let nanos = elapsed.as_secs() * 1_000_000_000 + elapsed.subsec_nanos() as u64;
+                self.client_id.observe(nanos);
+            }
+        }
+
+        pub fn stop_and_record(mut self) {
+            self.observe(true);
+        }
+
+        pub fn stop_and_discard(mut self) {
+            self.observe(false);
+        }
     }
 }